+++ core/src/routes/(editor)/editor/code.svelte
@@ -2,6 +2,7 @@
     import { globalComponentCollectionStore } from "../../globals/globalstores";
 </script>
 
-<div class="container-fluid">
+<div class="w-full">
     <pre>{JSON.stringify($globalComponentCollectionStore, null, 2)}</pre>
+
 </div>