+++ core/src/routes/menu/layers/layerstree.svelte
@@ -1,0 +1,36 @@
+<script>
+    export let component;
+    export let level = 0;
+</script>
+
+<div style:margin-left="20px"> <!-- {level * 20} -->
+    <slot>
+        {component}
+    </slot>
+
+    {#each component?.children ?? [] as child}
+    
+            <svelte:self component={child} level={level + 1}>
+                <span class="flex flex-row gap-2">
+                    <span class="align-middle"></span>{child.type} 
+                    <span class="flex flex-col">
+                        <span class="text-[8px] align-middle">{child.uuid}</span><br/> 
+                        <span class="text-[8px] align-middle">{child.data.class}</span> 
+                    </span> 
+                </span>
+            </svelte:self>
+        
+    {/each}
+</div>
+
+
+
+<!-- <div style:margin-left="{level * 20}px">
+    <slot {component} />
+
+    {#each component?.children ?? [] as child}
+        <svelte:self component={child} level={level + 1}>
+            <slot component={child} /> 
+        </svelte:self>
+    {/each}
+</div> -->